Remove summary page
